skills/podcast-visualizer/requirements.txt
# YouTube内容下载
yt-dlp>=2023.11.16

# 语音识别
openai-whisper>=20231117

# 说话人分离
pyannote.audio>=3.1.1
torch>=2.1.0
torchaudio>=2.1.0

# 后端API
fastapi>=0.104.0
uvicorn>=0.24.0

# 前端
streamlit>=1.28.0

# 数据处理
pydantic>=2.5.0
python-dotenv>=1.0.0

# 工具
tqdm>=4.66.0
pyyaml>=6.0
